get rid of deferred events completely
